pyproject.toml
[project]
name = "day21-ragas-demo"
version = "0.1.0"
description = "Collaborative RAG + RAGAS evaluation demo"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4",
    "chromadb>=0.5",
    "datasets",
    "fastapi",
    "httpx",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "loguru",
    "python-dotenv",
    "ragas>=0.2",
    "sentence-transformers>=3.0",
    "uvicorn",
]
